refactor: use proper layout groups for app router
