--- app/src/app/app/page.tsx
@@ -1,9 +1,0 @@
-const App = () => {
-    return (
-        <div>
-            app is working
-        </div>
-    )
-}
-
-export default App
