+++ app/src/components/footer/index.tsx
@@ -1,0 +1,9 @@
+export default function Footer() {
+            <footer className="mt-20 flex justify-between">
+            <div className="flex flex-col gap-2 font-medium capitalize tracking-wider">
+                <div className="  ">PRIVACY & LEGAL</div>
+                <div className="">CONTACT</div>
+            </div>
+            <div>© 2023 - One Day, All rights reserved.</div>
+        </footer>
+}
